git-svn-id: http://mc-server.googlecode.com/svn/trunk@504 0a769ca7-a7f5-676a-18bf-c427514a06d6

+++ source/FastNBT.h
@@ -259,7 +259,7 @@ protected:
 		{
 			// Compound: add the type and name:
 			m_Result.push_back((char)a_Type);
-			WriteString(a_Name.c_str(), a_Name.length());
+			WriteString(a_Name.c_str(), (short)a_Name.length());
 		}
 		else
 		{
